fix tests build

+++ marine/tests/wasm_tests/memory_limiting/src/effector.rs
@@ -16,6 +16,7 @@
 
 #![allow(improper_ctypes)]
 #![allow(clippy::all)]
+#![allow(non_snake_case)]
 
 use marine_rs_sdk::marine;
 
@@ -24,7 +25,7 @@ pub fn main() {}
 #[marine]
 pub fn allocate_single_module_single_piece(size: i64) -> u32 {
     let addr = Vec::with_capacity(size as usize).leak().as_ptr();
-    unsafe { std::mem::transmute::<*const u8, u32>(addr) }
+    unsafe { std::mem::transmute::<*const u8, usize>(addr) as u32}
 }
 
 #[marine]
@@ -34,7 +35,7 @@ pub fn allocate_single_module_1KB_pieces(mut size: i64) -> u32 {
     while size > 0 {
         unsafe {
             let addr = Box::leak(Box::new([0u8; 1024]));
-            acc ^= std::mem::transmute::<*const u8, u32>(addr.as_ptr());
+            acc ^= std::mem::transmute::<*const u8, usize>(addr.as_ptr()) as u32;
 
             size -= 1024
         }
